requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
google-generativeai>=0.8.0
google-api-python-client>=2.108.0
google-auth>=2.24.0
langdetect>=1.0.9
beautifulsoup4>=4.12.0
lxml>=4.9.0
httpx>=0.24.0
bleach>=6.0.0
markdown>=3.5.0
playwright>=1.40.0
slowapi>=0.1.9

# Testing
pytest-asyncio>=0.21.0

# Graphics library
openfigma @ git+https://github.com/federicodeponte/openfigma.git@main
